Conflicts:
	src/compiler/ast-misc.bas
	src/compiler/ast-node-const.bas
	src/compiler/ast-node-conv.bas
	src/compiler/ast-node-misc.bas
	src/compiler/ast.bi
	src/compiler/emit_x86.bas
	src/compiler/ir-hlc.bas
	src/compiler/parser-assignment.bas
	src/compiler/rtl-data.bas
	src/compiler/rtl-file.bas
	src/compiler/rtl-print.bas
	src/compiler/rtl-string.bas
	src/compiler/symb-data.bas
	src/compiler/symb-mangling.bas
	src/compiler/symb-struct.bas
	src/compiler/symb.bi

+++ examples/misc/CUnit/simple.bas
@@ -37,7 +37,7 @@ dim shared as FILE ptr temp_file = NULL
  ' Opens the temporary file used by the tests.
  ' Returns zero on success, non-zero otherwise.
  '/
-function init_suite1 cdecl() as integer
+function init_suite1 cdecl() as long
    temp_file = fopen("temp.txt", "w+")
    if temp_file = NULL then
       return -1
@@ -50,7 +50,7 @@ end function
  ' Closes the temporary file used by the tests.
  ' Returns zero on success, non-zero otherwise.
  '/
-function clean_suite1 cdecl() as integer
+function clean_suite1 cdecl() as long
    if fclose(temp_file) <> 0 then
       return -1
    else
