+++ src/gfxlib2/unix/fb_gfx_x11.h
@@ -58,7 +58,7 @@ extern int *fb_hX11FetchModes(int depth, int *size);
 extern int fb_hX11EnterFullscreen(int *h);
 extern void fb_hX11LeaveFullscreen(void);
 extern void fb_hX11InitWindow(int x, int y);
-extern int fb_hX11ScreenInfo(int *width, int *height, int *depth, int *refresh);
+extern int fb_hX11ScreenInfo(ssize_t *width, ssize_t *height, ssize_t *depth, ssize_t *refresh);
 
 #endif
