+++ src/rtlib/dev_file_encod_write.c
@@ -6,12 +6,11 @@ int fb_DevFileWriteEncod( FB_FILE *handle, const void* buffer, size_t chars )
 {
     FILE *fp;
     char *encod_buffer;
-    int bytes;
+	ssize_t bytes;
 
     FB_LOCK();
 
     fp = (FILE*) handle->opaque;
-
 	if( fp == NULL ) {
 		FB_UNLOCK();
 		return fb_ErrorSetNum( FB_RTERROR_ILLEGALFUNCTIONCALL );
@@ -28,7 +27,7 @@ int fb_DevFileWriteEncod( FB_FILE *handle, const void* buffer, size_t chars )
 	if( encod_buffer != NULL )
 	{
 		/* do write */
-		if( fwrite( encod_buffer, 1, bytes, fp ) != bytes )
+		if( fwrite( encod_buffer, 1, bytes, fp ) != (size_t)bytes )
 		{
 			FB_UNLOCK();
 			return fb_ErrorSetNum( FB_RTERROR_FILEIO );
