+++ src/rtlib/dos/sys_getexename.c
@@ -2,7 +2,7 @@
 
 #include "../fb.h"
 
-char *fb_hGetExeName( char *dst, int maxlen )
+char *fb_hGetExeName( char *dst, ssize_t maxlen )
 {
 	strncpy(dst, __fb_ctx.argv[0], maxlen);
 	char *p = strrchr( dst, '\\' );
