+++ src/rtlib/hook_lineinp.c
@@ -2,8 +2,15 @@
 
 #include "fb.h"
 
-FBCALL int fb_LineInput( FBSTRING *text, void *dst, int dst_len, int fillrem,
-                         int addquestion, int addnewline )
+FBCALL int fb_LineInput
+	(
+		FBSTRING *text,
+		void *dst,
+		ssize_t dst_len,
+		int fillrem,
+		int addquestion,
+		int addnewline
+	)
 {
     FB_LINEINPUTPROC lineinputproc;
